refactor(elysia): nest files/plugins routers — one file per endpoint
Restructure the flat routes-elysia/{files,plugins}.ts into nested trees
mirroring the test layout. Each endpoint gets its own file; shared
schemas and helpers live in model.ts.

src/routes-elysia/files/
  index.ts (21)         — aggregates sub-routers via .use()
  model.ts (42)         — TypeBox schemas for all endpoints
  file.ts (112)         — /api/file w/ path-traversal guard
  graph.ts, context.ts, read.ts, doc.ts, logs.ts
  plugins.ts, plugin-by-name.ts — legacy flat PLUGINS_DIR variants

src/routes-elysia/plugins/
  index.ts (8)          — canonical router
  model.ts (112)        — shared scanner (nested + flat + basename fallback)
  list.ts, get-by-name.ts

Largest file: 112 lines. No zod — TypeBox (built-in t.*) only. /api/file
uses t.String pattern as first-line traversal defense, with explicit
runtime re-check so the response stays the canonical 400 JSON shape.

Tested 16 Elysia smoke scenarios via .handle() — traversal guards,
wasm content-type, basename fallback, 404s all pass. All 20 existing
Hono contract tests still pass. bun run build passes.
