+++ src/routes-elysia/plugins/list.ts
@@ -1,0 +1,4 @@
+import { Elysia } from 'elysia';
+import { scanPlugins } from './model.ts';
+
+export const pluginsListRoute = new Elysia().get('/api/plugins', () => scanPlugins());
